requirements/development.txt
# Development pip packages.

-r base.txt

django-debug-toolbar

# Documentation
Sphinx
sphinx_rtd_theme

# Test coverage
coverage
flake8
wheel
twine
hatch
hatchling

# CLI
ipython
